Feature/thfhjs fix bug (#1021)
* Convert int type to double

* Import  textStyle

* Shield the 'TAB' key and modify the default drawing

* change config file

* change Config file

* Datagrid Cell ficusable set False

+++ App/TianHua.Hvac.UI/LoadCalculation/UI/AccidentAirConfig.xaml.cs
@@ -64,7 +64,7 @@ namespace TianHua.Hvac.UI.LoadCalculation.UI
             }
             else
             {
-                this._data.AirNum = int.Parse(AirNumTxt.Text);
+                this._data.AirNum = double.Parse(AirNumTxt.Text);
             }
             if (string.IsNullOrEmpty(TotalTxt.Text))
             {
@@ -72,7 +72,7 @@ namespace TianHua.Hvac.UI.LoadCalculation.UI
             }
             else
             {
-                this._data.TotalValue = int.Parse(TotalTxt.Text);
+                this._data.TotalValue = double.Parse(TotalTxt.Text);
             }
             this.Close();
         }
